yo mama 2

+++ Proj_2/Vector3.cpp
@@ -6,7 +6,7 @@ Vector3::Vector3(){
     _z = 0;
 }
 
-Vector3::Vector3(double x, double y, double z){
+Vector3::Vector3(float x, float y, float z){
 	_x = x;
     _y = y;
     _z = z;
@@ -16,33 +16,33 @@ Vector3::~Vector3(){
 	//nothing to do
 }
 
-double Vector3::getX(){
+float Vector3::getX(){
 	return _x;
 }
 
-double Vector3::getY(){
+float Vector3::getY(){
 	return _y;
 }
 
-double Vector3::getZ(){
+float Vector3::getZ(){
 	return _z;
 }
 
-void Vector3::set(double x, double y, double z){
+void Vector3::set(float x, float y, float z){
 	_x = x;
     _y = y;
     _z = z;
 }
 
-void Vector3::setX(double x) {
+void Vector3::setX(float x) {
 	_x = x;
 }
 
-void Vector3::setY(double y) {
+void Vector3::setY(float y) {
 	_y = y;
 }
 
-void Vector3::setZ(double z) {
+void Vector3::setZ(float z) {
 	_z = z;
 }
 
@@ -53,26 +53,26 @@ Vector3* Vector3::operator=(Vector3 &vec){
     return this;
 }
 
-Vector3* Vector3::operator*(double num){
-	double x = _x*num;
-    double y = _y*num;
-    double z = _z*num;
+Vector3* Vector3::operator*(float num){
+	float x = _x*num;
+    float y = _y*num;
+    float z = _z*num;
     Vector3* vec_aux = new Vector3(x, y, z);
     return vec_aux;
 }
 
 Vector3* Vector3::operator+(Vector3 &vec){
-	double x = _x + (&vec)->getX();
-    double y = _y + (&vec)->getY();
-    double z = _z + (&vec)->getZ();
+	float x = _x + (&vec)->getX();
+    float y = _y + (&vec)->getY();
+    float z = _z + (&vec)->getZ();
     Vector3* vec_aux = new Vector3(x, y, z);
     return vec_aux;
 }
 
 Vector3* Vector3::operator-(Vector3 &vec){
-	double x = _x - (&vec)->getX();
-    double y = _y - (&vec)->getY();
-    double z = _z - (&vec)->getZ();
+	float x = _x - (&vec)->getX();
+    float y = _y - (&vec)->getY();
+    float z = _z - (&vec)->getZ();
     Vector3* vec_aux = new Vector3(x, y, z);
     return vec_aux;
 }
