+++ Proj_2/Vector4.cpp
@@ -5,7 +5,7 @@ Vector4::Vector4(){
     Vector3();
 }
 
-Vector4::Vector4(double x, double y, double z, double w){
+Vector4::Vector4(float x, float y, float z, float w){
 	_w = w;
     _xyz = new Vector3(x, y, z);
 }
@@ -14,23 +14,23 @@ Vector4::~Vector4(){
 	delete _xyz;
 }
 
-double Vector4::getX(){
+float Vector4::getX(){
 	return _xyz->getX();
 }
 
-double Vector4::getY(){
+float Vector4::getY(){
 	return _xyz->getY();
 }
 
-double Vector4::getZ(){
+float Vector4::getZ(){
 	return _xyz->getZ();
 }
 
-double Vector4::getW(){
+float Vector4::getW(){
 	return _w;
 }
 
-void Vector4::set(double x, double y, double z, double w){
+void Vector4::set(float x, float y, float z, float w){
 	_w = w;
     _xyz->set(x, y, z);
 }
@@ -41,29 +41,29 @@ Vector4* Vector4::operator=(Vector4 &vec){
     return this;
 }
 
-Vector4* Vector4::operator*(double num){
-	double x = _xyz->getX() * num;
-    double y = _xyz->getY() * num;
-    double z = _xyz->getZ() * num;
-    double w = _w * num;
+Vector4* Vector4::operator*(float num){
+	float x = _xyz->getX() * num;
+    float y = _xyz->getY() * num;
+    float z = _xyz->getZ() * num;
+    float w = _w * num;
     Vector4* vec_aux = new Vector4(x, y, z, _w);
     return vec_aux;
 }
 
 Vector4* Vector4::operator+(Vector4 &vec){
-	double x = _xyz->getX() + (&vec)->getX();
-    double y = _xyz->getY() + (&vec)->getY();
-    double z = _xyz->getZ() + (&vec)->getZ();
-    double w = _w + (&vec)->getW();
+	float x = _xyz->getX() + (&vec)->getX();
+    float y = _xyz->getY() + (&vec)->getY();
+    float z = _xyz->getZ() + (&vec)->getZ();
+    float w = _w + (&vec)->getW();
     Vector4* vec_aux = new Vector4(x, y, z, _w);
     return vec_aux;
 }
 
 Vector4* Vector4::operator-(Vector4 &vec){
-	double x = _xyz->getX() - (&vec)->getX();
-    double y = _xyz->getY() - (&vec)->getY();
-    double z = _xyz->getZ() - (&vec)->getZ();
-    double w = _w - (&vec)->getW();
+	float x = _xyz->getX() - (&vec)->getX();
+    float y = _xyz->getY() - (&vec)->getY();
+    float z = _xyz->getZ() - (&vec)->getZ();
+    float w = _w - (&vec)->getW();
     Vector4* vec_aux = new Vector4(x, y, z, _w);
     return vec_aux;
 }
